refactor : HomeNavigation을 model 패키지로 이동 및 네비게이션 구조 개선
