+++ presentation/src/main/kotlin/com/moa/app/presentation/model/MoaSideEffect.kt
@@ -4,7 +4,7 @@ import androidx.navigation3.runtime.NavKey
 
 sealed interface MoaSideEffect {
     @JvmInline
-    value class Navigate(val destination: NavKey) : MoaSideEffect
+    value class Navigate(val destination: RootNavigation) : MoaSideEffect
 
     @JvmInline
     value class Dialog(val dialog: MoaDialogProperties?) : MoaSideEffect
