+++ presentation/src/main/kotlin/com/moa/app/presentation/ui/home/beforework/BeforeWorkViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.moa.app.core.model.onboarding.WorkPolicy
 import com.moa.app.presentation.bus.MoaSideEffectBus
 import com.moa.app.presentation.model.MoaSideEffect
-import com.moa.app.presentation.navigation.HomeNavigation
+import com.moa.app.presentation.model.HomeNavigation
 import com.moa.app.presentation.ui.home.beforework.model.BeforeWorkIntent
 import com.moa.app.presentation.ui.home.beforework.model.BeforeWorkUiState
 import dagger.assisted.Assisted
