+++ presentation/src/main/kotlin/com/moa/app/presentation/ui/home/working/WorkingScreen.kt
@@ -45,7 +45,7 @@ import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.moa.app.core.model.onboarding.Time
 import com.moa.app.presentation.R
-import com.moa.app.presentation.navigation.HomeNavigation
+import com.moa.app.presentation.model.HomeNavigation
 import com.moa.app.presentation.designsystem.component.MoaScheduleAdjustBottomSheet
 import com.moa.app.presentation.designsystem.component.MoaPrimaryButton
 import com.moa.app.presentation.designsystem.component.MoaTertiaryButton
